created a utils file for common utils.
