+++ uart_shell_arduino/src/cli/commands.h
@@ -1,9 +1,6 @@
 #ifndef COMMANDS_H
 #define COMMANDS_H
 
-#define MAX_ARGS 10
-#define MAX_PINS 13
-
 typedef void (*CommandFunction)(char* args);
 
 struct CommandEntry {
@@ -23,8 +20,5 @@ void cmd_version(char* args);
 void cmd_say_hello(char* args);
 void cmd_led(char* args);
 void cmd_set(char* args);
-
-//move later
-void str_to_upper(char* str);
 
 #endif
